third_party/WebKit/LayoutTests/http/tests/devtools/bindings/shadowdom-navigator.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script src="../../inspector/bindings/bindings-test.js"></script>

<script>

async function test() {
  var sourcesNavigator = new Sources.SourcesNavigatorView();
  sourcesNavigator.show(UI.inspectorView.element);

  TestRunner.markStep('dumpInitialNavigator');
  SourcesTestRunner.dumpNavigatorView(sourcesNavigator, false);

  TestRunner.markStep('attachShadow1');
  await Promise.all([
    BindingsTestRunner.attachShadowDOM('shadow1', '#template'),
    BindingsTestRunner.waitForSourceMap('sourcemap-script.js.map'),
    BindingsTestRunner.waitForSourceMap('sourcemap-style.css.map'),
  ]);
  SourcesTestRunner.dumpNavigatorView(sourcesNavigator, false);

  TestRunner.markStep('attachShadow2');
  await Promise.all([
    BindingsTestRunner.attachShadowDOM('shadow2', '#template'),
    BindingsTestRunner.waitForSourceMap('sourcemap-script.js.map'),
    BindingsTestRunner.waitForSourceMap('sourcemap-style.css.map'),
  ]);
  SourcesTestRunner.dumpNavigatorView(sourcesNavigator, false);

  TestRunner.markStep('detachShadow1');
  await Promise.all([
    BindingsTestRunner.detachShadowDOM('shadow1'),
    BindingsTestRunner.waitForStyleSheetRemoved('sourcemap-style.css'),
  ]);
  SourcesTestRunner.dumpNavigatorView(sourcesNavigator, false);

  TestRunner.markStep('detachShadow2');
  await Promise.all([
    BindingsTestRunner.detachShadowDOM('shadow2'),
    BindingsTestRunner.waitForStyleSheetRemoved('sourcemap-style.css'),
  ]);
  SourcesTestRunner.dumpNavigatorView(sourcesNavigator, false);

  TestRunner.completeTest();
}

</script>

</head>
<body onload="runTest()">
<p>
Verify that navigator shows proper UISourceCodes as shadow dom 
styles and scripts are added and removed.
</p>

<template id='template'>
<style>div {
    color: blue;
}
/*# sourceURL=sourcemap-style.css */
/*# sourceMappingURL=resources/sourcemap-style.css.map */
</style>
<script>window.foo = console.log.bind(console, 'foo');
//# sourceURL=sourcemap-script.js
//# sourceMappingURL=resources/sourcemap-script.js.map
</script>
<p>Hi! I'm ShadowDOM v1!</p>
</template>

</body>
</html>
