third_party/WebKit/LayoutTests/http/tests/devtools/bindings/shadowdom-navigator-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify that navigator shows proper UISourceCodes as shadow dom styles and scripts are added and removed.


Running: dumpInitialNavigator
top
  127.0.0.1:8000
    devtools/resources
      inspected-page.html

Running: attachShadow1
top
  127.0.0.1:8000
    devtools
      resources
        inspected-page.html
      bindings/resources
        sourcemap-typescript.ts
        sourcemap-sass.scss
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: attachShadow2
top
  127.0.0.1:8000
    devtools
      resources
        inspected-page.html
      bindings/resources
        sourcemap-typescript.ts
        sourcemap-sass.scss
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: detachShadow1
top
  127.0.0.1:8000
    devtools
      resources
        inspected-page.html
      bindings/resources
        sourcemap-typescript.ts
        sourcemap-sass.scss
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: detachShadow2
top
  127.0.0.1:8000
    devtools
      resources
        inspected-page.html
      bindings/resources
        sourcemap-typescript.ts
  (no domain)
    sourcemap-script.js
